third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-with-top-layer-008-expected.txt

This is a testharness.js-based test.
[FAIL] Contentvisibilityautostatechange fires once when c-v auto's dialog ancestor show()
  assert_equals: expected 1 but got 2
Harness: the test ran to completion.
